refactor(frontend): separate sdoc page entry renderers
