+++ frontend/config/webpack.entry.js
@@ -10,9 +10,10 @@ const entryFiles = {
   wiki2: '/wiki2.js',
   fileHistory: '/file-history.js',
   fileHistoryOld: '/file-history-old.js',
-  sdocFileHistory: '/pages/sdoc-file-history/index.js',
-  sdocPublishedRevision: '/pages/sdoc-published-revision/index.js',
-  sdocThumbnail: '/pages/sdoc-thumbnail',
+  sdocFileHistory: '/sdoc-file-history.js',
+  sdocPublishedRevision: '/sdoc-published-revision.js',
+  sdocThumbnail: '/sdoc-thumbnail.js',
+  sdocRevision: '/sdoc-revision.js',
   app: '/app.js',
   sharedDirView: '/shared-dir-view.js',
   sharedFileViewMarkdown: '/shared-file-view-markdown.js',
