+++ frontend/src/pages/sdoc-file-history/index.js
@@ -1,11 +1,8 @@
 import React from 'react';
-import { I18nextProvider } from 'react-i18next';
 import { UncontrolledTooltip } from 'reactstrap';
 import { DiffViewer } from '@seafile/seafile-sdoc-editor';
 import classnames from 'classnames';
 import dayjs from 'dayjs';
-import { createRoot } from 'react-dom/client';
-import i18n from '../../_i18n/i18n-sdoc-editor';
 import GoBack from '../../components/go-back';
 import Loading from '../../components/loading';
 import toaster from '../../components/toast';
@@ -476,9 +473,4 @@ class SdocFileHistory extends React.Component {
   }
 }
 
-const root = createRoot(document.getElementById('wrapper'));
-root.render(
-  <I18nextProvider i18n={ i18n } >
-    <SdocFileHistory />
-  </I18nextProvider>
-);
+export default SdocFileHistory;
